Display_message function change and important info
Display_message now has a width and height layout added to it so items don't overlap one another. Want to see if there's a way to make text not overlap in an easier way. Also, important info now displayed to user, like car get functions and inventory (needs to be formatted better)